+++ tests/unit/Exceptions/InvalidSearchOrderIdFilter.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace ild78\tests\unit\Exceptions;
+
+use atoum;
+use ild78;
+use Psr;
+
+class InvalidSearchOrderIdFilter extends atoum
+{
+    public function testClass()
+    {
+        $this
+            ->testedClass
+                ->extends(ild78\Exceptions\InvalidSearchFilter::class)
+        ;
+    }
+
+    public function testGetDefaultMessage()
+    {
+        $this
+            ->if($class = $this->testedClass->getClass())
+            ->then
+                ->string($class::getDefaultMessage())
+                    ->isIdenticalTo('Invalid order ID.')
+        ;
+    }
+
+    public function testGetLogLevel()
+    {
+        $this
+            ->if($class = $this->testedClass->getClass())
+            ->then
+                ->string($class::getLogLevel())
+                    ->isIdenticalTo(Psr\Log\logLevel::DEBUG)
+        ;
+    }
+}