+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First implementation
 - Tests on 7.1, 7.2 and 7.3
+- `Payment::list()` method
 
 ### Changes
 - `Object` renamed in `AbstractObject`